requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.1
python-dotenv==1.0.0
solana==0.31.0
solders==0.20.0
httpx==0.26.0
base58==2.1.1
pymongo==4.6.1
pydantic==2.5.2
python-multipart==0.0.6
